fix: rename timespan helper file
