+++ services/booking-api/src/helpers/timeSpanHelper.ts
